removed addfunctions2new in idfreader.py
